api/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
google_api_python_client==2.139.0
openai==1.41.0
python-dotenv==1.0.1
